phase-2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for the Todo application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "sqlmodel>=0.0.8",
    "asyncpg>=0.27.0",
    "python-dotenv>=0.19.0",
    "python-jose[cryptography]>=3.3.0",
    "alembic>=1.11.0",
]